bitkeeper revision 1.294 (3f0a9f16gdwvRs89xXMLxr3DvV1dgg)
Switch from using magic messages to map domain memory to using
a magic ioctl.

There's something very weird in this code: if userspace forgets to
unmap the memory, the next time it's called, it crashes.  It looks
like the old code had the same bug, so I'm going to pretend it isn't
there for a little while.