bitkeeper revision 1.1159.4.1 (4113d1b8oHof-8weGeQ3gNQFteLGEg)
David Becker writes:
This patch adds support for high memory to /proc/xen/memory_target,
and it restores the mem= linux boot parameter.  This is for 2.4.26 only.
I haven't looked at 2.6 yet.  (remmeber that CONFIG_HIGHMEM
is not on by default in the xen0 and xenU configs)
I have racks of 2g and 4g machines, hence my interest.

The 'mem=' param tells linux the maximum amount of memory it can use.
The feature originated for hosts where the normal mem detection mechanisms
failed.   These days its main use in stock linux is to artificially
constrain how much memory linux uses for testing low memory systems.

With this patch, mem= now means the max memory linux could ever use.
When a domain is created with less than that the mem= value, linux will
behave as though that 'missing' memory is 'allocated'.   To give the
domain more memory, first run setdomainmaxmem in Dom-0 to raise Xen's
limit, then write the new total to /proc/xen/memory_target in the
domain.   When mem= is not explicitly set, it defaults to the 
boottime size of the domain.
 dom-0# xm create name=dom-1 memory=100 extra='-b mem=2g'
 dom-0# setdomainmaxmem 1 2g
 dom-1# echo 2g > /proc/xen/memory_target