bitkeeper revision 1.1159.69.3 (4138d7a65FvXU3lh0Vx8Nsl4KhPxGw)
Fix potential security hole in writeable pagetable implementation: 
We wern't ensuring that that L1 pages' VA backpointer is immutable
after the backpointer is initialised  when the page first becomes 
linked into a pagetable. The backpointer can only be released after
the type count drops to zero (or 1 if the page is pinned).
In summary: We now ensure that if an L1 page is used in multiple 
pagetables it must be at the same virtual address in all of them,
and that L1 pages can only be used once in any given pagetable.
None of these extra rules should be a problem for any OS.