bitkeeper revision 1.1159.170.96 (41e8fbeayXDKhR25_xBQP3hTDrVJbQ)
Symptom:
When using a udev-based dom0, if you try to create a domU with a virtual
sda device and you do not have a local scsi device xm create fails.

Fix:
When using udev, only the device nodes needed are made.  The follow
patch changes Xend to return a device number based on:

http://www.lanana.org/docs/device-list/devices.txt

For sd and hd devices if stat() fails.  It was tested against both 2.0.2
and 2.0.3.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>