bitkeeper revision 1.1159.222.1 (41ebdd9btvi-aV_bkwfCgKSenC9XbQ)
Add an optional parameter (vcpus) to the xc_linux_build function replacing
the getenv() previously used and removing the requirement of using maxcpus
kernel parameter to limit the number of virtual cpus a guest uses.  The value
can now be controlled in the domain configuration files.

The default value of 1 is set in XenDomainInfo.py but is overridden by
parsing the config value.

Signed-off-by: Ryan Harper <ryanh@us.ibm.com>