bitkeeper revision 1.1159.212.45 (41f7969elyNTzsGsx57vqxgKK9SrRQ)
This change set moves the USB back and front end driver source code from
the 2.4 sparse tree to the 2.6 sparse tree in preparation for porting the
usb virtualization to 2.6.

This change set makes changes to the 2.4 mkbuildtree necessary to link
the files back into the 2.4 tree to preserve the previous usb
functionality for the 2.4 kernel.

The usbif.h file is also moved to be consistent with the location of the
netif.h file and necessary changes made to #include directives.
The usb code in the 2.6 tree is not yet integrated with the 2.6 build
process so the 2.6 build does not attempt to build it.

Tested with make world for both *2.4* and *2.6*.

Signed-off-by: butterwo@uk.ibm.com