bitkeeper revision 1.1159.223.83 (42097c41y12EFATQuIKPTDS1smnkWQ)
Subject: [Xen-devel] [PATCH] Fix initrd support for Linux 2.4/2.6

The following patch sets the default root device to be the null device. 
This is the behavior Linux is launched with no root device specified in
the command line and without a device provided by the boot loader.
Previously, the default root device was set to be /dev/ram0.  This broke
initrd support.  This change was made in 2.6 to because of a bug
reported on September 14th with dynamic ipconfig.  I've recreated his
problem and looked through the code and the source of the problem was
that his kernel command line was incorrect. He was missing the BOOTP
field in the config line requiring that there be an actual root device.
This worked with 2.4 but only because 2.4 was also broken.  His patch
just introduced that bug into 2.6.
I've tested this against the 2.6 and 2.4 trees in a recent copy of
xen-unstable.  I've tested initrd support and ipconfig support
(including dynamic ip support).  They both now work as expected.
If anyone needs an initrd for testing let me know.
Regards,
-- 
Anthony Liguori
Linux Technology Center (LTC) - IBM Austin
E-mail: aliguori@us.ibm.com
Phone: (512) 838-1208
Signed-off-by: Anthony Liguori
Signed-off-by: ian@xensource.com