bitkeeper revision 1.1159.261.2 (420e3014BpIA6NnJTdNQGkfIDMtzmQ)
Added an in-memory ring buffer to which serial console output
can be temporarily redirected.  Mode is toggled by the "c" key on
the Xen console.
Nice hack for printk() intensive debugging modes.
When switching back to serial output, the current contents of the
buffer are first dumped.  Buffer defaults to 128Kb, but size can
be set on the Xen boot command line.