bitkeeper revision 1.1308 (4260510bQdQyEJ4w-XdZSU5iSTB_Pg)
The attached patch apic.patch updates xen/arch/x86/apic.c so that it is based on linux 2.6.10 code.  A few changes made to linux required reworking functions in other files as well so I held off on those changes.  For example, setup_APIC_timer(void *) changed to setup_APIC_timer(unsigned int).  The change has no real bearing on the Xen code -- it's just a question of how close to linux one wants to be.  However, that change goes with a change to smp_call_function which would affect quite a few other files/functions so I left it as is.  Most of the other changes are trivial.
xen/ac_timer.h was included but not needed by apic.c so I removed it from the list of #includes.

The new apic.c includes two new files: asm/io_ports.h and asm/mach_apic.h.  Patches to create those files are included as well.