bitkeeper revision 1.1338.1.2 (42665facQKpiHhfw13qIq439AkffOw)
Here is a patch to get Xen running on a 32-way Unisys ES7000 system.
Please note that this patch alone is not sufficient to do this. You will
also need the clustered APIC patch that I will be submitting soon. But
since this patch was to a general area in Xen I decided to submit it
separately. I think this patch will be required to get Xen running on
any system that has more than 16 CPUs.
Signed-off-by: Aravindh Puthiyaparambil <aravindh.puthiyaparambil@unisys.com>