bitkeeper revision 1.1389.1.34 (42778a6aQ30Jj7E2TxzCF9ZtVpGMtw)
Move PCI device scanning to dom0. Enable ACPI in dom0. This should greatly
reduce the complexity of xen and move the complexity of dealing with
hardware bugs and workarounds etc to dom0.

The ioapic local apic (and hence all the vectors) are owned by the hypervisor.
Dom0 enables the ACPI interpreter, handles PCI and ACPI based interrupt 
routing. 

New hypercalls to assign vectors and for accessing the ioapic.

Functionality not yet provided:

o acpi=off to support machines with broken or no acpi support.
o support for driver domains 

Signed-off-by: Arun Sharma <arun.sharma@intel.com>
Signed-off-by: Asit Mallick <asit.k.mallick@intel.com>