bitkeeper revision 1.1678 (42a405d8raiThtAq9NRXuuFg6ur1qA)
Attached patch fixes FP initialization bug with x86-64 xenlinux.=20

With this patch, # of testcase failures with LTP is < 5 (out of 800),
almost same results as the native x86-64 Linux. The other two patches
for LTP are coming soon; with these, all the testcases will pass (except
the ones that fail on the native x86-64 Linux).

Signed-off-by: Jun Nakajima <jun.nakajima@intel.com>