1. Remove workaround patch 2. Update scripts/vif-bridge to not call 'brctl delif'
When a domU is shutdown/destroyed and the netif is destroyed, the
notify_call_chain triggered from unregister_netdevice() will trigger the
bridge event handler and which will call the proper code to remove the
device from the bridge.

I can't see any reason why brtcl delif should be called when taking out
a domain if the call chain will delete the interface from the bridge
when the vif is destroyed automatically.

Signed-off-by: Ryan Harper <ryanh@us.ibm.com>